Remove redundant casts

+++ src/freenet/support/io/PaddedBucket.java
@@ -130,7 +130,7 @@ public class PaddedBucket implements Bucket, Serializable {
         if(size < MIN_PADDED_SIZE) size = MIN_PADDED_SIZE;
         if(size == MIN_PADDED_SIZE) return size;
         long min = MIN_PADDED_SIZE;
-        long max = (long)MIN_PADDED_SIZE << 1;
+        long max = MIN_PADDED_SIZE << 1;
         while(true) {
             if(max < 0)
                 throw new Error("Impossible size: "+size+" - min="+min+", max="+max);
